requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
jinja2>=3.1.0
sqlalchemy>=2.0.0
python-multipart
httpx>=0.24.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
aiofiles>=23.0.0
openai>=1.0.0
pillow>=10.0.0
